Implement Delete book button

+++ app/javascript/components/DisplayBook.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import AppContext from '../contexts/AppContext'
 import { DELETE_BOOK_EVENT } from '../actions'
 import axios from 'axios'
-
+import { MdDeleteForever } from 'react-icons/md'
 
 const BookLists = styled.ul`
   display: flex;
@@ -18,6 +18,16 @@ const BookList = styled.li`
   height: 467px;
   padding: 0 10px 20px 10px;
 `
+const ImageTitleWrapper = styled.div`
+  padding-top: 3px;
+  padding-left: 5px;
+  padding-right: 5px;
+  display: flex;
+  justify-content: space-between;
+`
+const ImageTitle = styled.p`
+  font-size: 13px;
+`
 
 const BookImage = styled.img`
   height: 100%;
@@ -27,20 +37,26 @@ const BookImage = styled.img`
 function DisplayBook() {
   const {state: { showBooks }, dispatch } = useContext(AppContext);
   const handleClickBookDelete = (book) => {
-    axios.delete(`/api/v1/books/destroy/${book.id}`)
-    .then(() => {
-      dispatch({ type: DELETE_BOOK_EVENT, id: book.id })
-    })
-    .catch((e) => {
-      console.log(e)
-    })
+    const isDelete = window.confirm(`${book.title}を削除してよろしいですか？`)
+    if(isDelete) {
+      axios.delete(`/api/v1/books/destroy/${book.id}`)
+      .then(() => {
+        dispatch({ type: DELETE_BOOK_EVENT, id: book.id })
+      })
+      .catch((e) => {
+        console.log(e)
+      })
+    }
   }
-  console.log(showBooks)
   return (
     <BookLists>
       {showBooks.map((book, key) => (
-        <BookList key={key} onClick={() => handleClickBookDelete(book)}>
-          <BookImage src={book.url}/>{book.title}
+        <BookList key={key}>
+          <BookImage src={book.url}/>
+          <ImageTitleWrapper>
+            <ImageTitle>{book.title}</ImageTitle>
+            <MdDeleteForever onClick={() => handleClickBookDelete(book)} />
+          </ImageTitleWrapper>
         </BookList>
       ))}
     </BookLists>
